requirements.txt
python-pptx
xlsxwriter
xlrd
pygame
numpy
pillow
flask
flask-wtf
flask-restful
flask-sqlalchemy
PyQT5
PyQt6
pylint
pytest
python-telegram-bot
openpyxl
PyQt5-sip
PyQt6-sip
PyQtWebEngine
pyvirtualdisplay
pyautogui
python-docx
docxtpl
flake8
mrjob
pyspark
pandas
scikit-learn
seaborn
pymorphy3
pymorphy3-dicts-ru
networkx